// client/src/services/openapi/models/InterventiDettEdit.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type InterventiDettEdit = {
    ID: string;
    idIntervento: number;
    idTipoIntervento: string;
    riga: number;
    codiceTipoAttrezzatura: string | null;
    note: string | null;
    codiceMagazzino: string | null;
    codiceServizio: string | null;
    qta: number | null;
    prezzo: number | null;
    netto: number | null;
    utdc: string | null;
    utdv: string | null;
};