// client/src/services/openapi/models/InterventiMastBasic.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type InterventiMastBasic = {
    ID: number;
    numero: number | null;
    dataInserimento: string | null;
    dataEsecuzione: string | null;
    internoEsterno: string | null;
    tecnico: string | null;
    utcc: number | null;
    utcv: number | null;
    utdc: string | null;
    utdv: string | null;
    tipoRichiesta: string | null;
    idCliente: string | null;
    codiceArticolo: string | null;
    idMatricola: string | null;
    barcode: string | null;
    descrizione: string | null;
    note: string | null;
    idFornitore: string | null;
    nrDocFor: string | null;
    dataDocFor: string | null;
    codiceMagazzino: string | null;
    serialeDoc: string | null;
    matricolaFornitore: string | null;
    idDestinazione: string | null;
};